perlin_noise/requirements.txt
numpy
Pillow
perlin-noise